rename: 파스칼 케이스로 컴포넌트 명 변경
