+++ src/app/(auth)/sign-up/_components/email/EmailCheckWrapper.tsx
@@ -6,34 +6,22 @@ import { emailCheckSchema, emailCheckType } from "@/lib/zod/schema-email";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { useToast } from "@/utils/useToast";
-import PasswordForm from "./password-form";
-import {
-  getPasswordValidationErrors,
-  passwordCheckSchema,
-  passwordCheckType,
-} from "@/lib/zod/schema-password";
+import EmailForm from "./EmailForm";
 
-const PasswordCheckWrapper = () => {
+const EmailCheckWrapper = () => {
   const router = useRouter();
   const { toast } = useToast();
-  const formMethods = useForm<passwordCheckType>({
-    resolver: zodResolver(passwordCheckSchema),
+  const formMethods = useForm<emailCheckType>({
+    resolver: zodResolver(emailCheckSchema),
     mode: "onChange",
   });
 
   const {
-    watch,
     formState: { isValid },
   } = formMethods;
 
-  const password = watch("password");
-  const validation = getPasswordValidationErrors(password);
-
   const handleSubmit = () => {
     formMethods.handleSubmit(data => {
-      // console.log(data.password);
-      // console.log(data.passwordConfirm);
-
       // console.log("이메일" + data.email);
 
       /** api 호출 로직 추후 구현 */
@@ -42,20 +30,24 @@ const PasswordCheckWrapper = () => {
       //   variant: "error",
       //   description: `이미 가입된 이메일입니다.\n 다른 이메일로 가입해주세요.`,
       // });
-      router.push("/sign-up?state=USER_INFO");
+      router.push("/sign-up?state=PASSWORD");
     })();
   };
 
   return (
     <article className="flex flex-col justify-between min-h-real-screen pb-[56px]">
       <div className="pt-[139px] flex flex-col space-y-3">
-        <PasswordForm formMethods={formMethods} validation={validation} />
+        <p className="text-t4 text-grayscale-800">이메일 입력</p>
+        <EmailForm formMethods={formMethods} />
       </div>
-      <Button variant={isValid ? "active" : "deactive"} onClick={handleSubmit} disabled={!isValid}>
+      <Button
+        variant={isValid ? "active" : "deactive"}
+        onClick={handleSubmit}
+        disabled={!formMethods.formState.isValid}>
         다음
       </Button>
     </article>
   );
 };
 
-export default PasswordCheckWrapper;
+export default EmailCheckWrapper;
