+++ src/app/(auth)/sign-up/_components/email/EmailForm.tsx
