+++ src/app/(auth)/sign-up/_components/password/PasswordCheckWrapper.tsx
@@ -6,22 +6,34 @@ import { emailCheckSchema, emailCheckType } from "@/lib/zod/schema-email";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { useToast } from "@/utils/useToast";
-import EmailForm from "./email-form";
+import PasswordForm from "./PasswordForm";
+import {
+  getPasswordValidationErrors,
+  passwordCheckSchema,
+  passwordCheckType,
+} from "@/lib/zod/schema-password";
 
-const EmailCheckWrapper = () => {
+const PasswordCheckWrapper = () => {
   const router = useRouter();
   const { toast } = useToast();
-  const formMethods = useForm<emailCheckType>({
-    resolver: zodResolver(emailCheckSchema),
+  const formMethods = useForm<passwordCheckType>({
+    resolver: zodResolver(passwordCheckSchema),
     mode: "onChange",
   });
 
   const {
+    watch,
     formState: { isValid },
   } = formMethods;
 
+  const password = watch("password");
+  const validation = getPasswordValidationErrors(password);
+
   const handleSubmit = () => {
     formMethods.handleSubmit(data => {
+      // console.log(data.password);
+      // console.log(data.passwordConfirm);
+
       // console.log("이메일" + data.email);
 
       /** api 호출 로직 추후 구현 */
@@ -30,24 +42,20 @@ const EmailCheckWrapper = () => {
       //   variant: "error",
       //   description: `이미 가입된 이메일입니다.\n 다른 이메일로 가입해주세요.`,
       // });
-      router.push("/sign-up?state=PASSWORD");
+      router.push("/sign-up?state=USER_INFO");
     })();
   };
 
   return (
     <article className="flex flex-col justify-between min-h-real-screen pb-[56px]">
       <div className="pt-[139px] flex flex-col space-y-3">
-        <p className="text-t4 text-grayscale-800">이메일 입력</p>
-        <EmailForm formMethods={formMethods} />
+        <PasswordForm formMethods={formMethods} validation={validation} />
       </div>
-      <Button
-        variant={isValid ? "active" : "deactive"}
-        onClick={handleSubmit}
-        disabled={!formMethods.formState.isValid}>
+      <Button variant={isValid ? "active" : "deactive"} onClick={handleSubmit} disabled={!isValid}>
         다음
       </Button>
     </article>
   );
 };
 
-export default EmailCheckWrapper;
+export default PasswordCheckWrapper;
